+++ src/Contracts/OwnsEmizorCredentials.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Emizor\SDK\Contracts;
+
+interface OwnsEmizorCredentials
+{
+    public function getEmizorCredentials();
+}