+++ src/Contracts/RegisterContract.php
@@ -7,5 +7,5 @@ use Emizor\SDK\Models\BeiAccount;
 
 interface RegisterContract
 {
-    public function register(Closure $callback): BeiAccount;
+    public function register(Closure $callback): string;
 }