+++ src/DTO/RegisterDTO.php
@@ -11,7 +11,9 @@ class RegisterDTO
     public function __construct(
         public string $host,
         public string $clientId,
-        public string $clientSecret
+        public string $clientSecret,
+        public ?string $ownerId = null,
+        public ?string $ownerType = null
     )
     {
         $this->validate();
@@ -28,12 +30,16 @@ class RegisterDTO
             'host' => $this->host,
             'client_id' => $this->clientId,
             'client_secret' => $this->clientSecret,
+            'owner_id' => $this->ownerId,
+            'owner_type' => $this->ownerType,
         ];
 
         $rules = [
             'host' => ['required', 'string', 'in:PILOTO,PRODUCTION'],
             'client_id' => ['required', 'string', 'min:4', 'max:8'],
             'client_secret' => ['required', 'string', 'min:10'],
+            'owner_id' => ['nullable', 'string'],
+            'owner_type' => ['nullable', 'string', 'in:' . implode(',', array_keys(config('emizor_sdk.owners', [])))],
         ];
 
         $messages = [
@@ -47,6 +53,12 @@ class RegisterDTO
         ];
 
         $validator = Validator::make($data, $rules, $messages);
+
+        $validator->after(function ($validator) {
+            if (($this->ownerId && !$this->ownerType) || (!$this->ownerId && $this->ownerType)) {
+                $validator->errors()->add('owner', 'Ambos owner_id y owner_type deben estar presentes o ausentes.');
+            }
+        });
 
         if ($validator->fails()) {
             $errors = $validator->errors()->all();
@@ -63,7 +75,9 @@ class RegisterDTO
         return new self(
             $data['host'] ?? '',
             $data['client_id'] ?? '',
-            $data['client_secret'] ?? ''
+            $data['client_secret'] ?? '',
+            $data['owner_id'] ?? null,
+            $data['owner_type'] ?? null
         );
     }
 
@@ -73,6 +87,8 @@ class RegisterDTO
             'host' => $this->host,
             'client_id' => $this->clientId,
             'client_secret' => $this->clientSecret,
+            'owner_id' => $this->ownerId,
+            'owner_type' => $this->ownerType,
         ];
     }
 }